Most of the warnings in debug-mode are clarified

+++ src/shared/m_std.f90
@@ -75,7 +75,7 @@ contains
     real(DP), intent(inout), allocatable :: array(:)
     real(DP), intent(in) :: val
     !! --
-    integer, allocatable :: tmp(:)
+    real(DP), allocatable :: tmp(:)
     integer :: n
 
     n = size(array)
